fix: Force CosmosDB connection mode to Gateway
Every now and then, we are getting an issue where the Function App is not being triggered when we add new documents to CosmosDB.
This seems to be coming from the fact that the connection mode is unspecified (or set to Direct), which is causing the issue.
This fix forces the connection mode to Gateway, which fixed the issue in our tests.

noissue